+++ Backend/Entities/GrinderDial.cs
@@ -3,9 +3,9 @@ namespace Backend.Entities;
 public class GrinderDial : BaseEntity
 {
   public string Name { get; set; } = string.Empty;
-  public decimal? Min { get; set; }
-  public decimal? Max { get; set; }
-  public decimal? Step { get; set; }
+  public float? Min { get; set; }
+  public float? Max { get; set; }
+  public float? Step { get; set; }
   public string? Note { get; set; }
 
   public int GrinderId { get; set; }
